dom/media/test/test_mp4_pixel_aspect_ratio_hevc.html

<!DOCTYPE HTML>
<html>
<head>
<title>Test pixel aspect ratio for MP4 HEVC</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script type="text/javascript" src="manifest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
<script type="application/javascript">

/**
 * This test checks whether mp4 files are played backed correctly w.r.t.
 * display size, which will be determined by SAR information from either SPS or
 * PASP box. We prefer PASP box over SPS.
 */
add_task(async function testMP4PixelAspectRatio() {
  const tests = [
    {
      file: 'hevc_pasp_64_45_no_sar_in_sps.mp4',
      expectedWidth: 1024,
      expectedHeight: 576,
      pixelRatioFrom : "PASP Box"
    },
    {
      file: 'hevc_pasp_64_45_sps_sar_2.mp4',
      expectedWidth: 1024,
      expectedHeight: 576,
      pixelRatioFrom : "PASP Box"
    },
  ];
  for (let test of tests) {
    let video = document.createElement('video');
    info(`try to play ${test.file}, pixel aspect ratio from ${test.pixelRatioFrom}`);
    video.src = test.file;
    video.controls = true;
    document.body.appendChild(video);
    ok(await video.play().then(_=>true, _=>false), "video started playing");
    is(video.videoWidth, test.expectedWidth, `expect width ${test.expectedWidth}, actual width ${video.videoWidth}`);
    is(video.videoHeight, test.expectedHeight, `expect height ${test.expectedHeight}, actual height ${video.videoHeight}`);
    ok(await new Promise(r => video.onended = r).then(_=>true, _=>false),
       "video played to end");
    removeNodeAndSource(video);
  }
});

</script>
</head>
<body>
</body>
</html>
